backend/requirements.txt
django==3.1.5
djangorestframework==3.12.2
gunicorn==20.0.4
